refactored shapelets design and showed it works
