+++ autogalaxy/config/priors/light/linear/shapelets/cartesian.yaml
