+++ autogalaxy/config/priors/light/linear/shapelets/exponential.yaml
