+++ autogalaxy/config/priors/light/linear/shapelets/polar.yaml
