+++ autogalaxy/profiles/light/linear/__init__.py
@@ -6,3 +6,4 @@ from .exponential import Exponential, ExponentialSph
 from .dev_vaucouleurs import DevVaucouleurs, DevVaucouleursSph
 from .sersic_core import SersicCore
 from .exponential_core import ExponentialCore
+from .shapelets.polar import ShapeletPolar, ShapeletPolarEll
