+++ autogalaxy/profiles/light/linear/shapelets/polar.py
@@ -1,16 +1,11 @@
-import numpy as np
-from scipy.special import factorial, genlaguerre
-from typing import Optional, Tuple
+from typing import Tuple
 
-import autoarray as aa
+from autogalaxy.profiles.light import standard as lp
 
-from autogalaxy.profiles.light.decorators import (
-    check_operated_only,
-)
-from autogalaxy.profiles.light.shapelets.abstract import AbstractShapelet
+from autogalaxy.profiles.light.linear.abstract import LightProfileLinear
 
 
-class ShapeletPolarEll(AbstractShapelet):
+class ShapeletPolarEll(lp.shapelets.polar.ShapeletPolarEll, LightProfileLinear):
     def __init__(
         self,
         n: int,
@@ -40,65 +35,18 @@ class ShapeletPolarEll(AbstractShapelet):
             The (y,x) arc-second coordinates of the profile (shapelet) centre.
         ell_comps
             The first and second ellipticity components of the elliptical coordinate system.
+        intensity
+            Overall intensity normalisation of the light profile (units are dimensionless and derived from the data
+            the light profile's image is compared too, which is expected to be electrons per second).
         beta
             The characteristic length scale of the shapelet basis function, defined in arc-seconds.
         """
 
-        self.n = n
-        self.m = m
+        super().__init__(n=n, m=m, centre=centre, ell_comps=ell_comps, beta=beta, intensity=1.0)
 
-        super().__init__(centre=centre, ell_comps=ell_comps, beta=beta)
-
-    @aa.grid_dec.grid_2d_to_structure
-    @check_operated_only
-    @aa.grid_dec.transform
-    @aa.grid_dec.relocate_to_radial_minimum
-    def image_2d_from(
-        self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None
-    ) -> np.ndarray:
-        """
-        Returns the Polar Shapelet light profile's 2D image from a 2D grid of Polar (y,x) coordinates.
-
-        If the coordinates have not been transformed to the profile's geometry (e.g. translated to the
-        profile `centre`), this is performed automatically.
-
-        Parameters
-        ----------
-        grid
-            The 2D (y, x) coordinates in the original reference frame of the grid.
-
-        Returns
-        -------
-        image
-            The image of the Polar Shapelet evaluated at every (y,x) coordinate on the transformed grid.
-        """
-
-        radial = (grid[:, 0] ** 2 + grid[:, 1] ** 2) / self.beta**2.0
-        theta = np.arctan(grid[:, 1] / grid[:, 0])
-
-        laguerre = genlaguerre(n=(self.n - np.abs(self.m)) / 2.0, alpha=np.abs(self.m))
-
-        shapelet = laguerre(radial)
-
-        const = (
-            ((-1) ** ((self.n - np.abs(self.m)) / 2))
-            * np.sqrt(
-                factorial((self.n - np.abs(self.m)) / 2)
-                / factorial((self.n + np.abs(self.m)) / 2)
-            )
-            / self.beta
-            / np.sqrt(np.pi)
-        )
-        gauss = np.exp(-radial / 2.0)
-
-        return np.abs(
-            const
-            * radial ** (np.abs(self.m / 2.0))
-            * shapelet
-            * gauss
-            * np.exp(0.0 + 1j * -self.m * theta)
-        )
-
+    @property
+    def lp_cls(self):
+        return lp.ShapeletPolarEll
 
 class ShapeletPolar(ShapeletPolarEll):
     def __init__(
@@ -132,3 +80,7 @@ class ShapeletPolar(ShapeletPolarEll):
         """
 
         super().__init__(n=n, m=m, centre=centre, ell_comps=(0.0, 0.0), beta=beta)
+
+    @property
+    def lp_cls(self):
+        return lp.ShapeletPolar
