--- autogalaxy/profiles/light/shapelets/__init__.py
@@ -1,6 +1,0 @@
-from .cartesian import ShapeletCartesian
-from .cartesian import ShapeletCartesianEll
-from .polar import ShapeletPolar
-from .polar import ShapeletPolarEll
-from .exponential import ShapeletExponential
-from .exponential import ShapeletExponentialEll
