+++ autogalaxy/profiles/light/standard/shapelets/abstract.py
@@ -1,13 +1,14 @@
 from typing import Tuple
 
-from autogalaxy.profiles.light.linear.abstract import LightProfileLinear
+from autogalaxy.profiles.light.abstract import LightProfile
 
 
-class AbstractShapelet(LightProfileLinear):
+class AbstractShapelet(LightProfile):
     def __init__(
         self,
         centre: Tuple[float, float] = (0.0, 0.0),
         ell_comps: Tuple[float, float] = (0.0, 0.0),
+        intensity: float = 1.0,
         beta: float = 1.0,
     ):
         """
@@ -19,7 +20,7 @@ class AbstractShapelet(LightProfileLinear):
 
           https://arxiv.org/abs/astro-ph/0105178
 
-        Shapelets are are described in the context of strong lens modeling in:
+        Shapelets are described in the context of strong lens modeling in:
 
           https://ui.adsabs.harvard.edu/abs/2016MNRAS.457.3066T/abstract
 
@@ -35,4 +36,4 @@ class AbstractShapelet(LightProfileLinear):
 
         self.beta = beta
 
-        super().__init__(centre=centre, ell_comps=ell_comps, intensity=1.0)
+        super().__init__(centre=centre, ell_comps=ell_comps, intensity=intensity)
